matplotlibtool: hold plot points in double precision
Points were cast to float32 on every path into a plot: the loaders, the
transform engine and the overlay arithmetic. A float32 carries twenty
four bits of integer exactly, and the x axis is the sample counter,
which passes that eight seconds into a capture; beyond it neighbouring
samples rounded to one x, so the FFT saw zero spacing and every
measurement off a large capture sat on a quantised axis. Colour data
stays float32, being a gradient rather than a coordinate.

+++ matplotlibtool/ArrayFieldIntegration.py
@@ -174,7 +174,7 @@ class ArrayFieldIntegration:
             )
 
         info["data"] = data
-        x = data[info["x_field"]].astype(np.float32)
+        x = data[info["x_field"]].astype(np.float64)
         transform_params = info["properties"]["transform_params"]
         color_field = info["properties"].get("color_field")
         color_data = None
@@ -189,7 +189,7 @@ class ArrayFieldIntegration:
         for field, plot_index in manager.array_fields[array_index].items():
             if plot_index is None:
                 continue
-            points_xy = np.column_stack((x, data[field].astype(np.float32)))
+            points_xy = np.column_stack((x, data[field].astype(np.float64)))
             self.viewer.plot_manager.replace_plot_points(
                 plot_index,
                 self.viewer.transform_engine.apply_transform(
@@ -233,8 +233,8 @@ class ArrayFieldIntegration:
 
         points_xy = np.column_stack(
             (
-                data[x_field].astype(np.float32),
-                data[field_name].astype(np.float32),
+                data[x_field].astype(np.float64),
+                data[field_name].astype(np.float64),
             )
         )
